+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+ - [PR #44](https://github.com/projectapheleia/avl-axi/pull/44) Exclusives: EXOKAY incorrectly overwriting SLVERR/DECERR on read, and exclusivity granted despite an errored beat
+
 ## [v0.7.1] 0 2026-06-14
 
 ### Fixed